feat: Make setup_data cloneable (#115)
Current PLONK SNARK setup data and VK (the same per cycle) can't be used
by reference. setup_data is mutably changed & the structure is not
cloneable. This leaves us with a set of options:
1. compute the setup data for each proof (slow runtime)
2. compute it once, save it to disk and load it from disk (fast
implementation, not optimal; more of a hack)
3. refactor the code to use setup_data as reference (the "right" fix,
but very difficult & time-consuming refactoring)
4. make the setup data cloneable, compute it at start and clone it on
subsequent calls (pragmatic option)

4 was chosen and this PR contains the changes needed to make it happen
  (including some nice `Debug` derives)